requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
pytz==2018.9

